rename to cli_bootstrapper
