+++ tools/rescue/src/cli_bootstrapper.rs
