+++ tools/rescue/tests/rescue_cli_creates_blob.rs
@@ -1,7 +1,7 @@
 mod support;
 
 use libra_rescue::{
-    diem_db_bootstrapper::BootstrapOpts,
+    cli_bootstrapper::BootstrapOpts,
     rescue_cli::{RescueCli, Sub},
 };
 use libra_smoke_tests::libra_smoke::LibraSmoke;
