+++ tools/rescue/tests/rescue_cli_framework_upgrade_valid.rs
@@ -1,6 +1,6 @@
 use libra_framework::release::ReleaseTarget;
 use libra_rescue::{
-    diem_db_bootstrapper::BootstrapOpts,
+    cli_bootstrapper::BootstrapOpts,
     rescue_cli::{RescueCli, Sub},
 };
 use libra_smoke_tests::libra_smoke::LibraSmoke;
